+++ core/config.py
@@ -7,16 +7,20 @@ class ConfigManager:
     """Class that manages a cached configuration"""
 
     allowed_to_change_in_command = {
-        'status', 'log_channel_id', 'mention', 'disable_autoupdates', 'prefix',
-        'main_category_id', 'sent_emoji', 'blocked_emoji'
+        'activity_message', 'activity_type', 'log_channel_id', 'mention', 'disable_autoupdates', 'prefix',
+        'main_category_id', 'sent_emoji', 'blocked_emoji', 'thread_creation_response', 'twitch_url'
         }
     
     internal_keys = {
-        'token', 'snippets', 'aliases', 'owners', 'modmail_api_token',
-        'guild_id', 'modmail_guild_id', 'blocked'
+        'snippets', 'aliases', 'blocked', 'notification_squad', 'subscriptions'
+        }
+    
+    protected_keys = {
+        'token', 'owners', 'modmail_api_token', 'guild_id', 'modmail_guild_id', 
+        'mongo_uri', 'github_access_token', 'log_url'
         }
 
-    valid_keys = allowed_to_change_in_command.union(internal_keys)
+    valid_keys = allowed_to_change_in_command.union(internal_keys).union(protected_keys)
 
     def __init__(self, bot):
         self.bot = bot
@@ -32,8 +36,11 @@ class ConfigManager:
         data = {
             'snippets': {},
             'aliases': {},
-            'blocked': {}
+            'blocked': {},
+            'notification_squad': {},
+            'subscriptions': {}
         }
+
         try:
             data.update(json.load(open('config.json')))
         except FileNotFoundError:
@@ -42,8 +49,6 @@ class ConfigManager:
             data.update(os.environ)
             data = {k.lower(): v for k, v in data.items() if k.lower() in self.valid_keys}
             self.cache = data
-
-        self.bot.loop.create_task(self.refresh())
 
     async def update(self, data=None):
         """Updates the config with data from the cache"""